+++ source/exceeds_expectations/expect.d
@@ -42,7 +42,7 @@ public struct Expect(TReceived)
     {
         if (!completed)
         {
-            throw new InvalidExpectationException(
+            throw new InvalidExpectationError(
                 "`expect` was called but no assertion was made at " ~
                 filePath ~ "(" ~ line.to!string ~ "): \n\n" ~
                 formatCode(readText(filePath), line, 2) ~ "\n",
@@ -108,7 +108,7 @@ public struct Expect(TReceived)
         {
             if (
                 cast(FailingExpectationError) e ||
-                cast(InvalidExpectationException) e
+                cast(InvalidExpectationError) e
             )
             {
                 throw e;
@@ -128,7 +128,7 @@ public struct Expect(TReceived)
 
         if (predicates.length == 0)
         {
-            throw new InvalidExpectationException(
+            throw new InvalidExpectationError(
                 "Missing predicates at " ~ filePath ~ "(" ~ line.to!string ~ "): \n" ~
                 "\n" ~ formatCode(readText(filePath), line, 2) ~ "\n",
                 filePath, line
@@ -208,7 +208,7 @@ public struct Expect(TReceived)
 
         if (predicates.length == 0)
         {
-            throw new InvalidExpectationException(
+            throw new InvalidExpectationError(
                 "Missing predicates at " ~ filePath ~ "(" ~ line.to!string ~ "): \n" ~
                 "\n" ~ formatCode(readText(filePath), line, 2) ~ "\n",
                 filePath, line
@@ -422,7 +422,7 @@ public struct Expect(TReceived)
         }
         catch (RegexException e)
         {
-            throw new InvalidExpectationException(
+            throw new InvalidExpectationError(
                 "toMatch received an invalid regular expression pattern at " ~
                 filePath ~ "(" ~ line.to!string ~ "): \n\n" ~
                 formatCode(readText(filePath), line, 2) ~ "\n" ~